Create Order implemtation changes

+++ order-service/src/main/java/com/bookstore/orders/clients/catalog/Product.java
@@ -1,11 +1,13 @@
 package com.bookstore.orders.clients.catalog;
 
+import java.math.BigDecimal;
+
 public record Product(
         String code,
         String name,
         String description,
         String imageUrl,
-        double price
+        BigDecimal price
 ) {
 }
